Adds defines for all(current) 133 spells, replaces clunky typepaths in jobs for adding spells with the defines (#2375)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Please read the title.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Turns typepaths that go sometimes 9 or 10 deep into simple defines that
can be dropped in and out.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)



https://github.com/user-attachments/assets/311f5128-6c13-4a28-b795-3789d99f3ab2



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->