Remove the seemingly arbitrary do_after for looking up or down, lowers the cooldown for looking up or down (#2370)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I really don't know what to say here that isn't captured by the title.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The current method of looking up and down is clunky and janky for
seemingly no reason, and this makes it more responsive and intuitive to
use. You're tilting your head slightly, not astral projecting, it
shouldn't take that much effort or time.

Down the line, if it's absolutely imperative that looking at a different
elevation is a pain in the ass, maybe we can add some behavior to heavy
armor helmets to slow it down. But for now, I think we should free the
people's necks and let them look up and down as they please.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)



https://github.com/user-attachments/assets/db5f3bca-3d75-4df6-9ddb-c9fac7b76620



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->