Proper bandit categories (#2353)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Removes "stuff" and "things", replaces them with, gear, apparel,
consumables, also moves the horse to knight and renames it to saiga,
since it is a saiga and only they get riding skill.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Stuff and things are awful names for categories.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/0fee39eb-883d-4283-8c05-38412870b53a)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->