Reverts new scroll text (#2397)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Gyran, while making the docks, decided to go ahead and change how scroll
text was read, this led to a text that was pixelated, low quality,
unreadable sometimes, and broke the moment you would write a slightly
long message. This simply reverts those changes and leaves the scrolls
back how like they were, so they'll read the same as a parchment, but
with a little overlay behind.

Here's a pretty decent explanation:
https://github.com/Rotwood-Vale/Ratwood-Keep/issues/2061
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Most people liked the text before this change, and a lot complained when
it was added. So hopefully this'll make people use scrolls for messages
again, instead of having to search for a parchment since they were just
objectively better for messages.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/00494a81-38b2-4a21-a16a-c10b69fdf1cc)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->