Fixes needle repairs for armor (#2407)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Repairing cloth items with a needle no longer causes a runtime that
prevents the repairer from, well, repairing.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I fixa da bug.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/6292b687-0a25-422f-bc92-5e90875572fa)



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->