Fixes "say laws" command on SCOM (#2398)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

While creating jabberlines, the creator probably accidentally deleted
the "state laws" and "say laws" command that let people that can't read
know the laws of the land. This fixes it!

Issue that let me know it:
https://github.com/Rotwood-Vale/Ratwood-Keep/issues/2185 (Thanks for
reporting it)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Bug fix!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/2b91a2e3-6a70-42c6-b332-62339cc56534)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->