Adds "Bad Eyesight" to Flaw List (#2412)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

See Title

## Why It's Good For The Game

Someone removed steward's personal flaw because they're too judgemental
to get around town, so now the Steward's bad-eyesight was left around to
rot. Some characters actually roleplay, so it should be nice to have
such a flaw enabled.

Flaws are not meant to be completely game-hindering, moreso a tool to
give users more character when they read Manga, Doujinshi, and
fanfiction to create them.

## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/268128ec-c3ec-4daf-9e0c-a05e4858ef0e)