Enables Bad Eyesight for selection/save (#2413)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Fixes #2412 by enabling it to be saved and properly selected.

When you originally selected "badsight" it would auto default to
"random" for some reason. Whoever did this, clearly wanted to avoid
someone adding this flaw in without some form of investigation needed.

## Why It's Good For The Game

See original PR


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/b3795676-668e-4a0b-898f-92d099ab158a)

Forgot how to squash, sorry.