Heraldry for weapons! (Skin changes) (#2411)
## About The Pull Request

This PR aims to give players more freedom to choose how their weapons
will look in their hands, this is still very basic, but i do pretend to
expand it to cover halberds, clubs and more skins for the ones already
existing in the future. (CHANGES ARE COSMETIC ONLY AND DO NOT GRANT ANY
ADVANTAGE OR DISADVANTAGE TO THE USER OR PEOPLE ON THE RECEIVING END OF
THE WEAPON)

Edit: All sprites already existed in the game files, i merely selected a
few i barely see/don't see at all during game play, so they are actually
used for something in the game

## Why It's Good For The Game

It's cosmetics! Everyone loves cosmetics

## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/9c790b5b-415e-41b6-ab52-c9c567f6f916)

![1212](https://github.com/user-attachments/assets/95738f90-af93-4f40-a4b5-27fea0385dfe)

![1313](https://github.com/user-attachments/assets/cecd1cab-0cd5-469a-bad7-32e13193c03f)

![image](https://github.com/user-attachments/assets/4772213a-51dd-405f-8f26-a9f12daa4d71)