fixes and minor changes.
Fixes silver goblets not being pickupable
Adjusts the name of chests from leech chests, to chests in wiz tower,
adds subtype of void dragon for event use,
Fixes dragon emotes runtiming and failing.
Tweaks cost of Tier one summons to be less of a unnessicary hassle. T1s are meant to be easy to summon