Adds warning message to a dry cloth. (#2416)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Currently the cloth doesn't tell you when it is dry. This fixes that.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
If you don't know the cloth is dry you just keep wiping until you
realize. Bit annoying.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/bc2d9bc1-3ddd-4ed3-88f6-c8b79093628d


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->