Liches bleed to death often (#2430)
## About The Pull Request

Fixes #1857 hopefully, regarding the question of liches bleeding to
death and having blood at all (they are skellies)

Basically, this adds a new trait to the game called NO_BLOOD which is
applied to liches only. This will change a few things regarding blood in
general as well as the heart_hud we have.

The heart is used to track your blood volume, and since liches do not
have blood we will be using **LIMB DAMAGE** instead to track how hurt
you are. It has a few other changes too in regards to text, but nothing
too important to mention here!!!

## Why It's Good For The Game

I don't want my skeleton evil wizards to bleed to death or bleed at all,
in fact i just removed their blood entirely.

## Proof of Testing (Required)

my gyazo is not working so you will have to trust me on this

![image](https://github.com/user-attachments/assets/2ca87ccd-e4e3-47c6-9a71-1a80fa5e310a)

![image](https://github.com/user-attachments/assets/567eb8bf-4351-47f3-8c91-7c65cde8a2ff)

![image](https://github.com/user-attachments/assets/18a9d9fd-2157-4578-9879-02b89a05cc9d)