make small benches passable from all dirs (#2428)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes it so you can walk over the small bench. Makes the most sense plus
it prevents the guessing game of knowing where the 'back' of it is.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
You should be able to walk over these things it's messed up that you
can't!!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/5313820c-74e6-4640-83ed-02639917f3af

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->