Allows broken armour to be repaired (#2425)
## About The Pull Request

This PR allows damaged armour to be repaired from 0% durability.

## Why It's Good For The Game

The intent of this PR is to ease the burden from smith players- at
present, any damaged armour has to be resmelted, and if it costs more
than one bar, is going to require new steel. This makes it so that every
time someone gets into a fight and has some armour broken, a smith has
to have them all over again- keeping them stuck at the smithy through
most of the round. Discussed this with Eora in the Discord, and decided
to try it out.


## Proof of Testing (Required)

https://gyazo.com/135d78d67a4a764cbe62957cedbd266b

Successful compile, tested ingame and the armour can be repaired and
loses the damaged appearance upon repair.