Soups and stews have less nutrients but more volume (#2423)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Before, while making my food PR, i lowered the amount of stew and soup
so they match up with the amount of nutriments they gave, so a single
cut onion wouldn't give 10x more food when boiled instead of cooked, but
the nutrition values were so high, that you needed such things as two
steaks to make a full bowl of stew. And said bowl of stew had the
equivalent of a full-course meal worth of food.

This PR simply lowers the amount of nutrients given by stews, soups, and
oatmeal but raises the amount of volume given, so a bowl of soup or a
bowl of stew will be easier to make, but they won't be as filling
anymore.

For comparison's sake, i'll give some examples on how much a full bowl
of stew and a full bowl of soup was equal to:
(I mean this in terms of how much nutrition they gave, not as in the
amount of ingredients required, but they should be about the same)
-------Old------- 
1 bowl of stew = 2 steaks cooked, or 1 full spiced and roasted chicken,
or 1 entire loaf of raisin bread
1 bowl of oatmeal = 1 meal (onion & steak, frybird & tato, wiener &
cabbage, wiener & onion, etc.)
1 bowl of soup = 1 cooked steak, or 1 wiener, or 1 whole salumoi, or 1
set of coppiettes
------New------
1 bowl of stew = same as old soup
1 bowl of oatmeal = 1 pie slice, or 1 cake slice, 1 frybird, 1 fried egg
1 bowl of soup = 1 sandwich (not cackleberry), 1 & 1/2 berries or
apples, 2 salumoi slices

This basically means that soups, stews, and oatmeal will become,
probably as they should be, cheaper & easier meals to make, but less
filling than most actual foods. While still giving you bonus nutrients
compared to plain cooked foods. Pretty much the poor man's meal.

Oh and, i also raised the taste multiplier for them, it doesn't fix that
the first sip always tastes like water, but after the second sip and any
subsequent ones the taste works correctly. So 90% of the soup will taste
like soup, and only 10% like water.

Amount that's needed now:
2 mince = 1 bowl of thick stew
1 oat = 1 & 1/2 bowls of oatmeal
1 vegetable (3 slices) = 1 & 1/2 bowls of soup.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Soups, stews, and oatmeal will be a cheap kind of food, probably as it
should've been on the first place. So now you can serve full bowls of
stew without having to water them down for them to be cheap, or worry
about not having enough ingredients to make a single bowl's worth.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/e04dff18-2cd7-41bc-b92b-73edd36480e9)

![image](https://github.com/user-attachments/assets/bd2cc2b3-6bb4-4a28-b437-42d0191522b2)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->