Descriptions for amulets (#2343)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR adds some descriptions to the amulets (Thank neonicon on the
discord! He's the one that wrote it) that should hopefully add a little
bit of flavor and lore to the game.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More lore! More flavor! More stuff for the acolytes to know about their
gods.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/8e49aa3b-02b3-4107-9e0c-9931ac383e7d)
I checked the descriptions in-game too, but forgot to take a screenshot,
sorry :/
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->