Fixes watchman helmet onmob sprite (#2437)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The clothing PR used the incorrect path for the onmob sprites of the
watchman helmet, which meant it didn't have any sprites, and looked like
this:

![image](https://github.com/user-attachments/assets/edd3e16c-2aff-404f-a779-3ef29c75f36e)
This fixes it, and makes it look like this again:

![image](https://github.com/user-attachments/assets/743cb4cb-a6d7-47ca-ba61-c182bb45ee80)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bug fix!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
The two pictures above, and: 

![image](https://github.com/user-attachments/assets/669fbbb7-d93f-4e42-ace5-3aaf4d3f35bb)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->