Fixes unearthing stumps, changed to use shovel (#2442)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes the stump removal code. 

Previously it seemed to go via the 'attacked_by' proc which only calls
during combat mode. This makes it use 'attackby' which does not require
combat mode activated. It also ports the code from Vanderlin using the
shovel instead. (currently always gives stump even if burnt) Works with
all shovel sub types as well.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
2 people and I tried to remove stumps for about 15 minutes not
understanding how it was done. When I looked at the code I noticed the
issue. I think not only does this make the most sense for removing a
stump but it also aligns with the method in other servers meaning it'll
be easier to transfer knowledge and also keep some familiarity in the
system. It also fixes the need to use combat mode. While it no longer
requires combat mode it still does work while in it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/990e1768-7bc7-4ad9-9846-6f8235951872

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->