Runelock damage fix (#2360)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Runelock damfactor wasn't being applied to bullets like other firearms.
This fixes that, bumping the damage from 40 -> 200, with 50 AP. This
puts it slightly below an Arquebus against someone in plate armor, and
above it to lesser armored targets. If this is too much it can be
adjusted, I am simply fixing the bug that was causing damfactor to not
apply.
## Why It's Good For The Game
Bugs bad.


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/547d0cb0-e882-4a78-b7fd-8101b029625f)

Co-authored-by: Enite <AA@AA>