Changes back the feed intent on containers to intent 3, Save my soul (#2455)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

due to this pr
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2394
feed intent was moved to intent 2 from 3
as a cause, this caused everyone who is used to pressing 3 to feed
people due to playing on codebases other than ratwood as well

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

it hurts my soul everytime someone about to die opens a health potion,
tries to drink it, then splashes it onto themselves and bleeds out
because someone swapped around splash and feed

please save my soul i beg

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/2a489d20-31c3-4cbc-99f5-17361e7fa865)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->