[COSMETIC] Joinlate changes (#2451)
## Pull Request Changelog
This is recreation of PR #2409, for a same reason as PR 2449 and 2450. 
We got 4 upvotes in 2409, this changes are demanded, probably?
- Job order in `jobs.dm` was changed, meaning results in joinlate
window.
   - Court magos and physician are moved to courtier lists.
   - Councilors role is moved to courtier lists.
   - Shuffled order in other role lists.
- Job .dm files of the physician and the magos were moved from
`nobility` folder to `courtier` folder.
- Mayor and town sheriff got `selection_color` with their appropriate
position value.

## Why It's Good For The Game
As continuation of PR 2363.
PR 2176 changed court physicians and magos' statuses to nobles - which
resulted for joinlate list changes, leaving the butler and the jester to
be only classes for the entire list on courtiers.

Looking at this disproportion, it was decided to move councilors, court
physician and court magos classes to courtier list. The nobility,
granted in PR 2176 as a trait feature, is NOT taken away. Arguments are
stored below spoiler segment:
<details> 
<summary>Open Spoiler</summary>
<ol>
<li>Major argument is, of course, the class name. They are Court
physician and magos and this makes them courtiers.
<li>The courtier status means no offence to nobility status, as there
are Bankrupt Aristocrat subclass for servants as well as Page subclass
for squires. Both are carrying the trait of nobility.
<li>In the preferences/class choice menu the court magos and the court
physician are divided from noble roles of the duke, the consort, heirs;
and knights are placed in garrison lists as well. This is leading us to
a concept, that noble list is listing command roles of the Rockhill,
similar to command roles for Space Station builds. The councilors are
placed right below Marshall for utility purposes (Who is subordinate to
whom), I guess, same as knight (Noble role in joinlate list) is placed
above squire in garrison list.
<li>Councilor argument: class tutorial suggests, that your character may
have bought his position, not earned like knight or obtain as heritage.
They are still noble educated courtiers, working for Royal Marshall.
Moreover: their path in code are
<code>jobs\job_types\roguetown\COURTIER\councillor.dm</code>.
</ol>
</details>

This PR would result in retaining of nobility for the magos, the
physician, the councilors, but cosmetic change for hierarchy lists, as
the physician and the magos normally are not ordering around (Except
servants and mage apprentices); and councilors are helping Royal
Marshall (Thus can't be even approximately equalized with powerholding
nobles), nothing beyond this obligations.
Remained noble trait would keep this roles tax-free in Nervelock
operations.

Probably it's worth to change "Nobles" title to "The Court"? 

Besides this, jobs order was shuffled in other joinlate lists, you can
take a look below. Main goal was to ensure command roles such as mayor
to be on top of the list, secondary - to put subordinates under their
bosses (This logic was already applied to nobles and churchemen)

## Proof of Testing (Required)
Looks much more proportionate, isn't it? 

![image](https://github.com/user-attachments/assets/bd10fee1-66f1-405a-9c3b-31b84f842e2a)