Underdwellers! (#2390)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Re-adds underdwellers from
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2303 with better
balance?
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Shunned races content good!! Especially for the most shunned and
forgotten of them all, kobolds and verminfolk.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/25153ee9-2432-4f57-b4ce-cdb97216cd80)

![image](https://github.com/user-attachments/assets/9fdd69e4-5277-437c-9546-4814aa020793)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->