Admin Additions - Up and Down arrows (#2467)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Replaces 'pai' icon for "up" and "down" admin arrows with proper arrows.
(Uses cursed Astrata symbol and set it up as arrows)

No one but admins would really see this. This is just makes it so you
don't click pai icon wondering if it is up or down.


![image](https://github.com/user-attachments/assets/a8672e57-54d1-4293-8033-8c5a563831fe)