Fireman changes (#2474)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes fireman (and hard grabbed people) to soft grab when they are
moved up a staircase
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People did not like fireman carried people to default to hard grab when
carrying people up stairs.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/7a32f4c2-0c07-4de8-b3ac-32f5a6e08421



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->