Confessor gear (#2464)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Add in a outfit from the server Twilight Fortress for the confessor.
The confessor starts off with the gear as well as a bronze mask, while
the inquisitor keeps all their unique gear but gets a silver mask in
their satchel in case they want to wear the mask as well.
The jacket is a very slightly buffed gambeson, with a max integrity of
150

## Why It's Good For The Game

Confessor deserves a new outfit, the inquisitor got new gear, and the
confessor still just gets a black leather vest

## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/b2441756-4ae1-4264-a6e6-578b8764d412)

![image](https://github.com/user-attachments/assets/f5795c78-16a8-4fc4-a1d5-679f8d8db43f)

![image](https://github.com/user-attachments/assets/56865e28-a282-4356-8c30-da0c0b0a2d92)

![image](https://github.com/user-attachments/assets/b2de73a4-0f00-4432-8095-301eef1b7d03)