Possible crash fix (#2487)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Reverting the code for the beard growth PR #2172, which is suspected to
be causing the server crashes, possibly causing a memory leak. This is
only suspected not guaranteed

Edit: While testing this code out on a private, it started causing
massive memory leaks, Dream seeker had started using close to 7 gigs of
ram after attempting to shave. It works now that i have removed all the
traces of the beard growth PR that i can.

Also i seem to have broken it so shaving doesn't work at all now, but
that is a project for later if this does fix the crashing, as im sure
people rarely use the shaving feature anyway

## Why It's Good For The Game
Crashes are bad


## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

![image](https://github.com/user-attachments/assets/fb593a22-4d2f-43bd-b306-ff3b086c075f)

![image](https://github.com/user-attachments/assets/968b00ba-d9a8-4611-82ca-a415a36454d8)