Longshoreman and Harbormaster racelock lifted (#2462)
## About The Pull Request

This PR changes the races of the retired **PIRATE CREW** from being
axian, lizard, and humen only to being very shunned and up.

## Why It's Good For The Game


As it stands the roles are former pirates. All historical record of
pirates shows them taking pretty much anyone with a heartbeat and
functioning limbs. They often freed slaves of any race and so forth for
recruiting.
Additionally, the roles are **dead**. I think most players don't even
know it exists. It's actually a cool concept that's just utterly
neglected as it stands, so hopefully we'll see something come of this
change.

I am considering just unchaining them from the merchant's authority,
since they have no functionality with him now in this new map change,
not that they had much to begin with besides being map-linked. Open to
suggestions.


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/50265b77-2a4e-444f-a720-68cd3641f975)