Adds back Cheese Makers (#2443)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! --> Adds a Cheese Maker towner role, as requested in
"https://discord.com/channels/1240735983276654733/1340759024840343642"

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->New towner
content, starts with a wheel of cheddar cheese, 3 bottles of milk, a
cloth (to seethe the cheese, idk how cheese crafting actually works but
from the code this is what is needed) and a skillet (not for cheesing).
Also gets a mid pouch to afford buying from the farm if needed and to
hire someone to build/buy materials for a cheese store


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/46091c32-15de-4efb-a35a-96254c662c57)

![image](https://github.com/user-attachments/assets/80198620-6687-4131-b36c-ff7f778ff5d5)

![image](https://github.com/user-attachments/assets/7a87d96a-1822-4b50-b117-5b82291faa4a)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->