Removes prob chance to fail cleaning floors (Why???) (#2508)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes the probability check that makes you fail cleaning turfs
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It never made sense to begin with, there's no message to tell you it
failed and there's no cleaning skill associated with it anyway. Until
there is it's such a simple task it should always succeed, no? Anyone
who brings it up gets mad.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/85661608-bf3e-4ed7-847a-b508eebbd3b9


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->