Makes brick walls climbable (#2509)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

The stone brick and decorated wall were spread all across the town in:
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2236

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The walls mentioned were used as decorative walls and with how prevalent
they are they are limiting the possibilities for climbing and
exploration quite a bit and the author of that pr agreed they should be
climbable.

![image](https://github.com/user-attachments/assets/222f565b-783b-472b-a38c-4f20afedb985)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/4732cb81-68ec-40d6-815a-19fcd0ed6ba6)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->