Arquebus pistol now can be used one-handed (#2485)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Does as the title says, it was probably overlooked by the creator for
the firearm rework, so this simply makes the arequebus pistol work
exactly as it did before, but it can't be wielded anymore and it can
shoot while one-handed.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bug fix! Pistol is now a one handed weapon, as it should be.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/acd2646b-5708-4fb2-9bac-356358d39a9b)

![image](https://github.com/user-attachments/assets/43ff8229-d92a-43c4-bbd1-612ae02f031e)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->